Code refactore for building gem
